+++ src/index.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import { HashRouter } from 'react-router-dom';
+import { HashRouter as Router } from 'react-router-dom';
 
 import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
@@ -7,7 +7,7 @@ import '@fortawesome/fontawesome-free/css/all.css';
 import { App } from './App';
 
 createRoot(document.getElementById('root') as HTMLDivElement).render(
-  <HashRouter>
+  <Router>
     <App />
-  </HashRouter>,
+  </Router>,
 );
